// src/app/enums/covid.enum.ts
export enum sortOrder{
    asc = "ASC",
    desc = "DESC"
}
export enum ColumnName {
    country = "Country",
    totalConfirmed = "Total Confirmed",
    newConfirmed = "New Confirmed",
    totalDeath = "Total Death",
    newDeath = "New Death",
    totalRecoverd = "Total Recovered",
    newRecovered = "New Recovered"
}